+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix variables support for gradle 7.4 (#34)
+
 ### 2.3.0 (2022-04-02)
 * Update packages:
   - mkdocs 1.2.3 -> 1.3.0 (fixes #29: jinja2 3.1.0 support)